introduce a ModuleDefinition class
