+++ example/add.dart
@@ -1,15 +1,14 @@
 import 'dart:io';
 
-import 'package:wasmi/execution.dart';
-import 'package:wasmi/format.dart';
+import 'package:wasmi/execute.dart';
+import 'package:wasmi/parse.dart';
 
 void main(List<String> args) {
-  var module = Module.parse(File('example/add.wasm'));
-  var function = module.exportedFunction('add')!.func as DefinedFunction;
-  var context = ExecutionContect(module);
+  var moduleDefinition = ModuleDefinition.parse(File('example/add.wasm'));
+  var module = Module(moduleDefinition);
 
   for (var i = 0; i <= 10; i++) {
-    final result = context.execute(function, [i, i]);
+    final result = module.call('add', [i, i]);
     print('add($i, $i) = $result');
   }
 }
