+++ example/fib.dart
@@ -1,15 +1,14 @@
 import 'dart:io';
 
-import 'package:wasmi/execution.dart';
-import 'package:wasmi/format.dart';
+import 'package:wasmi/execute.dart';
+import 'package:wasmi/parse.dart';
 
 void main(List<String> args) {
-  var module = Module.parse(File('example/fib.wasm'));
-  var context = ExecutionContect(module);
-  var function = module.exportedFunction('fib')!.func as DefinedFunction;
+  var moduleDefinition = ModuleDefinition.parse(File('example/fib.wasm'));
+  var module = Module(moduleDefinition);
 
   for (var i = 0; i <= 10; i++) {
-    final result = context.execute(function, [i]);
+    final result = module.call('fib', [i]);
     print('fib($i) = $result');
   }
 }
