+++ samples/gcd.dart
@@ -1,16 +1,15 @@
 import 'dart:io';
 
-import 'package:wasmi/execution.dart';
-import 'package:wasmi/format.dart';
+import 'package:wasmi/execute.dart';
+import 'package:wasmi/parse.dart';
 
 void main(List<String> args) {
-  var module = Module.parse(File('samples/gcd.wasm'));
-  var context = ExecutionContect(module);
-  var function = module.exportedFunction('gcd')!.func as DefinedFunction;
+  var moduleDefinition = ModuleDefinition.parse(File('samples/gcd.wasm'));
+  var module = Module(moduleDefinition);
 
   final arga = 24;
   final argb = 6;
 
-  final result = context.execute(function, [arga, argb]);
+  final result = module.call('gcd', [arga, argb]);
   print('gcd($arga, $argb) = $result');
 }
